+++ GeneralsMD/Code/GameEngine/Source/GameClient/GUI/GUICallbacks/Menus/WOLBuddyOverlay.cpp
@@ -249,8 +249,8 @@ WindowMsgHandledType BuddyControlSystem( GameWindow *window, UnsignedInt msg,
 					if(rc->pos < 0)
 						break;
 
-					GPProfile profileID = (GPProfile)GadgetListBoxGetItemData(control, rc->pos, 0);
-					RCItemType itemType = (RCItemType)(Int)GadgetListBoxGetItemData(control, rc->pos, 1);
+					GPProfile profileID = (GPProfile)(uintptr_t)GadgetListBoxGetItemData(control, rc->pos, 0);
+					RCItemType itemType = (RCItemType)(Int)(uintptr_t)GadgetListBoxGetItemData(control, rc->pos, 1);
 					UnicodeString nick = UnicodeString(GadgetListBoxGetText(control, rc->pos).str() + 2); // Skip the online/offline indicator
 
 					GadgetListBoxSetSelected(control, rc->pos);
@@ -302,7 +302,7 @@ WindowMsgHandledType BuddyControlSystem( GameWindow *window, UnsignedInt msg,
 				if (selected >= 0)
 				{
 #if defined(GENERALS_ONLINE)
-					int profileID = (int)GadgetListBoxGetItemData(buddyControls.listboxBuddies, selected);
+					int profileID = (int)(uintptr_t)GadgetListBoxGetItemData(buddyControls.listboxBuddies, selected);
 
 					// Block chatting with a buddy who is currently in the same game as us
 					if (TheNGMPGame && TheNGMPGame->isGameInProgress())
@@ -336,7 +336,7 @@ WindowMsgHandledType BuddyControlSystem( GameWindow *window, UnsignedInt msg,
 					}
 						
 #else
-					GPProfile selectedProfile = (GPProfile)GadgetListBoxGetItemData(buddyControls.listboxBuddies, selected);
+					GPProfile selectedProfile = (GPProfile)(uintptr_t)GadgetListBoxGetItemData(buddyControls.listboxBuddies, selected);
 					BuddyInfoMap *m = TheGameSpyInfo->getBuddyMap();
 					BuddyInfoMap::iterator recipIt = m->find(selectedProfile);
 					if (recipIt == m->end())
@@ -504,7 +504,7 @@ void updateBuddyInfo( void )
 
             GadgetListBoxGetSelected(buddyControls.listboxBuddies, &selected);
             if (selected >= 0)
-                selectedProfile = (GPProfile)GadgetListBoxGetItemData(buddyControls.listboxBuddies, selected);
+                selectedProfile = (GPProfile)(uintptr_t)GadgetListBoxGetItemData(buddyControls.listboxBuddies, selected);
 
             selected = -1;
             GadgetListBoxReset(buddyControls.listboxBuddies);
@@ -704,7 +704,7 @@ void updateBuddyInfo( void )
 
 	GadgetListBoxGetSelected(buddyControls.listboxBuddies, &selected);
 	if (selected >= 0)
-		selectedProfile = (GPProfile)GadgetListBoxGetItemData(buddyControls.listboxBuddies, selected);
+		selectedProfile = (GPProfile)(uintptr_t)GadgetListBoxGetItemData(buddyControls.listboxBuddies, selected);
 
 	selected = -1;
 	GadgetListBoxReset(buddyControls.listboxBuddies);
@@ -1075,7 +1075,7 @@ void WOLBuddyOverlayInit( WindowLayout *layout, void *userData )
 				GadgetListBoxGetSelected(buddyControls.listboxBuddies, &selected);
 				if (selected >= 0)
 				{
-					GPProfile profileID = (GPProfile)GadgetListBoxGetItemData(buddyControls.listboxBuddies, selected);
+					GPProfile profileID = (GPProfile)(uintptr_t)GadgetListBoxGetItemData(buddyControls.listboxBuddies, selected);
 
 
 					// sending to them, or getting from them, is valid
@@ -1278,7 +1278,7 @@ WindowMsgHandledType WOLBuddyOverlaySystem( GameWindow *window, UnsignedInt msg,
 				GadgetListBoxGetSelected(buddyControls.listboxBuddies, &selected);
 				if (selected >= 0)
 				{
-					GPProfile profileID = (GPProfile)GadgetListBoxGetItemData(control, selected);
+					GPProfile profileID = (GPProfile)(uintptr_t)GadgetListBoxGetItemData(control, selected);
 					UnicodeString nick = GadgetListBoxGetText(control, selected);
 
 					NGMP_OnlineServices_SocialInterface* pSocialInterface = NGMP_OnlineServicesManager::GetInterface<NGMP_OnlineServices_SocialInterface>();
@@ -1346,7 +1346,7 @@ WindowMsgHandledType WOLBuddyOverlaySystem( GameWindow *window, UnsignedInt msg,
 					if (rc->pos < 0)
 						break;
 
-					GPProfile profileID = (GPProfile)GadgetListBoxGetItemData(control, rc->pos);
+					GPProfile profileID = (GPProfile)(uintptr_t)GadgetListBoxGetItemData(control, rc->pos);
 					UnicodeString nick = GadgetListBoxGetText(control, rc->pos);
 
                     NGMP_OnlineServices_SocialInterface* pSocialInterface = NGMP_OnlineServicesManager::GetInterface<NGMP_OnlineServices_SocialInterface>();
@@ -1387,7 +1387,7 @@ WindowMsgHandledType WOLBuddyOverlaySystem( GameWindow *window, UnsignedInt msg,
 						break;
 
 					Bool isBuddy = false, isRequest = false;
-					GPProfile profileID = (GPProfile)GadgetListBoxGetItemData(control, rc->pos);
+					GPProfile profileID = (GPProfile)(uintptr_t)GadgetListBoxGetItemData(control, rc->pos);
 					UnicodeString nick = GadgetListBoxGetText(control, rc->pos);
 					BuddyInfoMap *buddies = TheGameSpyInfo->getBuddyMap();
 					BuddyInfoMap::iterator bIt;
@@ -1478,7 +1478,7 @@ WindowMsgHandledType WOLBuddyOverlaySystem( GameWindow *window, UnsignedInt msg,
 
 					GadgetListBoxGetSelected(listbox, &selected);
 					if (selected >= 0)
-						selectedName = TheNameKeyGenerator->keyToName((NameKeyType)(int)GadgetListBoxGetItemData(listbox, selected));
+						selectedName = TheNameKeyGenerator->keyToName((NameKeyType)(int)(uintptr_t)GadgetListBoxGetItemData(listbox, selected));
 
 					if (!selectedName.isEmpty())
 					{
